Python3 script to read all files in a directory line by line and use the application CLI utility 'toilet' to transform each line into colorful ascii art and print it to stdout.
This application is portable as it is writtern in python3.
But it was only tested on Xubuntu in application 'Xfce Terminal'.
Probably this script will run succesfully also on Windows and many other operating systems but probably the default terminal emulators on Windows and other systems will not be able to interpret the output of this script correctly and so probably it will be not colorful. But likely there are terminal emulators available for Windows and other operating systems that will be able to interpret the output of this script correctly.